==> scripts/06-models.R <==
#### Preamble ####
# Purpose: Logistic Regression
# Author: Michaela Drouillard
# Data: 17 March 2023
# Contact: michaela.drouillard@mail.utoronto.ca
# License: MIT

library(tidyr)
library(dplyr)
library(tidyverse)
library(lubridate)
library(tidymodels)
library(randomForest)
library(ranger)
library(caret)
library(themis)
library(yardstick)

data <- read_csv(here::here("inputs/data/clean_df.csv"))



data_reduced <- 
  data |> 
  select(producer,danceability, energy, loudness, mode, speechiness, acousticness, instrumentalness, liveness, valence,
         tempo
  ) 

write_csv(data_reduced,"inputs/data/data_reduced.csv")


#### Predicting Antonoff: non-bayesian  ####
## glmnet requires all variables to be numeric, so i'm excluding the producer column
set.seed(853)
# 1. Create a new binary column
antdata <- read_csv(here::here("inputs/data/antonoff_viz.csv"))
adf <- antdata |> 
  select(producer,danceability, energy, loudness, mode, speechiness, acousticness, instrumentalness, liveness, valence,
         tempo
  ) 
adf$is_antonoff <- ifelse(adf$producer == "antonoff", 1, 0)
adf$is_antonoff <- as.factor(adf$is_antonoff)
adf <- adf |>
  select(is_antonoff, danceability, energy, loudness, mode, speechiness, acousticness, instrumentalness, liveness, valence,
              tempo
) 


#write_csv(lr_data, "inputs/data/lr_data.csv")
adf$is_antonoff <- factor(adf$is_antonoff, levels = c("1", "0"))
# 2. Update variable names and strata
antonoff_split <- 
  adf |>
  initial_split(strata = is_antonoff)

antonoff_train <- training(antonoff_split)
antonoff_test <- testing(antonoff_split)
antonoff_folds <- vfold_cv(antonoff_train, strata = is_antonoff)


# 3. Update the recipe
is_antonoff <- 
  recipe(is_antonoff ~ ., data = antonoff_train) 


# we can `prep()` just to check that it works:
prep(is_antonoff)




glmnet_spec <- 
  logistic_reg(penalty = tune(), mixture = 1) |>
  set_engine("glmnet")

# Comparing the basic model and the downsampled one
wf_set <-
  workflow_set(
    list(basic = is_antonoff,
         downsampling = is_antonoff |> step_downsample(is_antonoff)),
    list(glmnet = glmnet_spec)
  )

wf_set

narrower_penalty <- penalty(range = c(-3, 0))

# install.packages("doParallel")
doParallel::registerDoParallel()

tune_rs <- 
  workflow_map(
    wf_set,
    "tune_grid",
    resamples = antonoff_folds,
    grid = 15,
    metrics = metric_set(sens, spec, accuracy, mn_log_loss),
    param_info = parameters(narrower_penalty)
  )

tune_rs

autoplot(tune_rs) + theme(legend.position = "none")

rank_results(tune_rs, rank_metric = "sens")
rank_results(tune_rs, rank_metric = "mn_log_loss")
rank_results(tune_rs, rank_metric = "accuracy")
rank_results(tune_rs, rank_metric = "spec")


## Downsampling vs Basic

downsample_rs <-
  tune_rs %>%
  extract_workflow_set_result("downsampling_glmnet")

# Tuning result for recipe that included downsampling + glmnet model
plotdownsample <- autoplot(downsample_rs) + ggtitle("Downsampled Data")

basic_rs <-
  tune_rs %>%
  extract_workflow_set_result("basic_glmnet")

plotbasic <- autoplot(basic_rs) + ggtitle("Original Data")

library(gridExtra)
grid.arrange(plotdownsample, plotbasic, ncol=2)

best_penalty <- 
  downsample_rs %>%
  select_by_one_std_err(-penalty, metric = "mn_log_loss")

best_penalty

# Last fit is a helper function that fits one time to the training data
# and evaluates one time on the testing data
final_fit <-  
  wf_set %>% 
  extract_workflow("downsampling_glmnet") %>%
  finalize_workflow(best_penalty) %>%
  last_fit(antonoff_split)

final_fit

collect_metrics(final_fit)

antonoffmatrix <- collect_predictions(final_fit) %>%
  conf_mat(is_antonoff, .pred_class)

antonoffmatrix



antonoffmatrix_df <- as.data.frame(antonoffmatrix$table)



antonoff_predictions <- collect_predictions(final_fit)

antonoff_predictions$is_antonoff <- factor(antonoff_predictions$is_antonoff, levels = c("1", "0"))
antonoff_predictions$pred_class <- factor(antonoff_predictions$.pred_class, levels = c("1", "0"))

str(antonoff_predictions$is_antonoff)
str(antonoff_predictions$pred_class)

conf_matrix <- confusionMatrix(antonoff_predictions$is_antonoff, antonoff_predictions$.pred_class)

precision_value <- conf_matrix$byClass['Pos Pred Value']
recall_value <- conf_matrix$byClass['Sensitivity']

# Print Precision and Recall
print(precision_value)
print(recall_value)



# install.packages("vip")
library(vip)
antonoff_vip <-
  extract_fit_engine(final_fit) %>%
  vi()

antonoff_vip
write_csv(antonoff_vip, "inputs/data/antonoff_vip.csv")

## Predictors ##


antonoff_vip %>%
  group_by(Sign) %>%
  slice_max(Importance, n = 15) %>%
  ungroup() %>%
  ggplot(aes(Importance, fct_reorder(Variable, Importance), fill = Sign)) + 
  geom_col() +
  facet_wrap(vars(Sign), scales = "free_y") +
  labs(y = NULL) +
  theme(legend.position = "none")


### SECOND MODEL: Bayesian logistic regression ###

library(ROSE)
library(rstanarm)

lr_data <- data_reduced
lr_data$is_antonoff <- ifelse(lr_data$producer == "antonoff", 1, 0)
lr_data$is_antonoff <- as.factor(lr_data$is_antonoff)


antonoff_split <- 
  lr_data |>
  initial_split(strata = is_antonoff)

antonoff_train <- training(antonoff_split)
antonoff_test <- testing(antonoff_split)

naive_prior <- normal(0, 5) #wide normal distribution approximation to flat prior



library(caret)
library(rstanarm)

antonoff_train$is_antonoff <- factor(antonoff_train$is_antonoff, levels = c("0", "1"))

ds_data<- downSample(x = antonoff_train[, !(names(antonoff_train) %in% "is_antonoff")],
                                y = antonoff_train$is_antonoff)



if("producer" %in% names(ds_data)) {ds_data <- ds_data[, !(names(ds_data) %in% "producer")]}

antonoff_train$is_antonoff <- factor(antonoff_train$is_antonoff)

antonoff_train <- antonoff_train[, !(names(antonoff_train) %in% "producer")]

bayesian_model <- stan_glm(
  is_antonoff ~ .,
  data = antonoff_train,
  family = binomial(link = "logit"),
  prior = naive_prior,  
  prior_intercept = normal(0, 5),  
  iter = 2000,  
  seed = 853  
)


summary(bayesian_model)





posterior_predictive <- posterior_predict(bayesian_model, newdata = antonoff_test)

predicted_probabilities <- apply(posterior_predictive, 2, mean)

predicted_classes <- ifelse(predicted_probabilities > 0.5, "1", "0")

predicted_classes <- factor(predicted_classes, levels = levels(antonoff_test$is_antonoff))

confusionMatrix(predicted_classes, antonoff_test$is_antonoff)

lr_data 


## UpSample 

library(caret)
library(rstanarm)
library(rsample)


antonoff_split <- initial_split(lr_data, strata = is_antonoff)
antonoff_train <- training(antonoff_split)
antonoff_test <- testing(antonoff_split)


naive_prior <- normal(0, 10) 

antonoff_train$is_antonoff <- factor(antonoff_train$is_antonoff, levels = c("0", "1"))


up_data <- upSample(x = antonoff_train[, !(names(antonoff_train) %in% c("is_antonoff", "producer"))],
                    y = antonoff_train$is_antonoff)



bayesian_model2 <- stan_glm(
  Class ~ .,
  data = up_data,
  family = binomial(link = "logit"),
  prior = naive_prior,  
  prior_intercept = normal(0, 5),  
  iter = 2000,  
  seed = 853  
)

summary(bayesian_model2)
posterior_predictive2 <- posterior_predict(bayesian_model2, newdata = antonoff_test)
predicted_probabilities2 <- apply(posterior_predictive2, 2, mean)
predicted_classes2 <- ifelse(predicted_probabilities2 > 0.5, "1", "0")
predicted_classes2 <- factor(predicted_classes2, levels = levels(antonoff_test$is_antonoff))
confusionMatrix(predicted_classes2, antonoff_test$is_antonoff)







### THIRD MODEL: Random Forest ###
# one hot encoding for each producer (is_antonoff 1,0, is_elwroth 1,0, etc)
# multi class classification
install.packages("ranger")
install.packages("caret") # For some helper functions
library(ranger)
library(caret)


data_reduced$producer <- as.factor(data_reduced$producer)


set.seed(853)
antonoff_split <- initial_split(data_reduced, strata = producer)
antonoff_train <- training(antonoff_split)
antonoff_test <- testing(antonoff_split)

rf_model <- ranger(producer ~ ., data = antonoff_train, importance = 'impurity')
#impurity == asking model to interpret variable importance based on gini impurity

rf_predictions <- predict(rf_model, data = antonoff_test)$predictions

confusionMatrix(rf_predictions, antonoff_test$producer)

## Variable Importance
var_importance <- importance(rf_model)
print(var_importance)
sorted_importance <- var_importance[order(-var_importance)]
print(sorted_importance)

## Variable importance for each producer

list_importance <- list()
unique_producers <- unique(data_reduced$producer)

for (producer in unique_producers) {
  data_temp <- data_reduced
  data_temp$producer_binary <- as.factor(ifelse(data_temp$producer == producer, 1, 0))
  
  rf_temp <- randomForest(producer_binary ~ . - producer, data = data_temp, ntree = 100)
  
  importance_temp <- rf_temp$importance
  print(colnames(importance_temp))  # To see available columns
  
  list_importance[[producer]] <- importance_temp[,'MeanDecreaseGini']
}

df_importance <- bind_rows(list_importance, .id = "producer")

head(df_importance)

df_importance_long <- df_importance %>%
  gather(variable, importance, -producer)

write.csv(df_importance, file = "inputs/data/df_importance.csv", row.names = FALSE)


ggplot(df_importance_long, aes(x = reorder(variable, importance), y = importance)) +
  geom_bar(stat = "identity") +
  facet_wrap(~producer, scales = "fixed") + 
  coord_flip() +
  labs(title = "Variable Importance by Producer", 
       x = "Variable", 
       y = "Mean Decrease Gini") +
  theme_minimal()



### Correlations by Producer ###


library(tidyr)
library(corrr)

# Assuming data_reduced is your dataset
correlations_by_producer <- data_reduced %>%
  group_by(producer) %>%
  nest() %>%
  mutate(correlation = map(data, ~ correlate(.x, use = "pairwise.complete.obs"))) %>%
  select(producer, correlation)

library(ggplot2)
library(reshape2) 

producer_name <- "antonoff"

correlation_matrix <- correlations_by_producer %>%
  filter(producer == producer_name) %>%
  pull(correlation) %>%
  .[[1]]

melted_corr <- melt(correlation_matrix)

ggplot(melted_corr, aes(x = term, y = variable, fill = value)) +
  geom_tile() +
  scale_fill_viridis_c() +
  theme_minimal() +
  labs(title = paste("Correlation Heatmap for", producer_name), fill = "Correlation")

##Correlations for each producer : Understanding style###


correlations_by_producer <- data_reduced %>%
  group_by(producer) %>%
  nest() %>%
  mutate(correlation = map(data, ~ correlate(.x, use = "pairwise.complete.obs"))) %>%
  select(producer, correlation)

# Melting all correlation matrices for all producers
melted_corr_list <- correlations_by_producer %>%
  mutate(melted = map(correlation, melt)) %>%
  select(producer, melted)

melted_corr_all <- melted_corr_list %>%
  unnest(cols = melted)


common_order <- c("tempo", "valence", "liveness", "instrumentalness", "acousticness", 
                  "speechiness", "mode", "loudness", "energy", "danceability")

# Make sure they are factors with the common order for both x and y axes
melted_corr_all$term <- factor(melted_corr_all$term, levels = common_order)
melted_corr_all$variable <- factor(melted_corr_all$variable, levels = common_order)

ggplot(melted_corr_all, aes(x = term, y = variable, fill = value)) +
  geom_tile() +
  scale_fill_viridis_c() +
  theme_minimal() +
  labs(title = "Correlation Heatmap", fill = "Correlation") +
  facet_wrap(~ producer, scales = "free", ncol = 2)


library(tidyverse)

table_data <- melted_corr_all %>%
  filter(as.numeric(term) < as.numeric(variable)) %>% 
  select(-term) %>%
  spread(key = variable, value = value)

write_csv(table_data, "inputs/data/table_data.csv")

producer_specific_table <- function(producer_name) {
  table_data %>% filter(producer == producer_name)
}


print(producer_specific_table("antonoff"))

problematic_cols <- melted_corr_all %>%
  filter(as.numeric(term) < as.numeric(variable)) %>%
  group_by(producer, variable) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  filter(count > 1)

print(unique(problematic_cols))

deduplicated_data <- melted_corr_all %>%
  filter(as.numeric(term) < as.numeric(variable)) %>%
  group_by(producer, term, variable) %>%
  slice(1) %>%
  ungroup()

table_data <- deduplicated_data %>%
  spread(key = variable, value = value)


## Range of correlation scores

library(dplyr)

range_data <- table_data %>%
  gather(variable, value, -producer, -term) %>%
  group_by(variable) %>%
  summarize(
    min_value = min(value, na.rm = TRUE),
    max_value = max(value, na.rm = TRUE),
    range = max_value - min_value
  ) %>%
  arrange(-range)


##Variables with greatest variance##
top_vars <- range_data %>% 
  top_n(3, range) %>%
  pull(variable)

print(top_vars)



selected_data <- table_data %>%
  select(producer, term, all_of(top_vars)) %>%
  arrange(producer, term)

write_csv(selected_data, "inputs/data/selected_data.csv")

print(selected_data)

library(ggplot2)

# Danceability Plot
ggplot(selected_data, aes(x=term, y=danceability, fill=producer)) +
  geom_bar(stat="identity", position="dodge") +
  labs(title="Danceability across Producers", y="Danceability Score") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

# Energy Plot
ggplot(selected_data, aes(x=term, y=energy, fill=producer)) +
  geom_bar(stat="identity", position="dodge") +
  labs(title="Energy across Producers", y="Energy Score") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

# Loudness Plot
ggplot(selected_data, aes(x=term, y=loudness, fill=producer)) +
  geom_bar(stat="identity", position="dodge") +
  labs(title="Loudness across Producers", y="Loudness Score") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

## Danceability Scores, descriptive ##

library(dplyr)

# Computing average danceability scores for each producer
danceability_summary <- data_reduced %>%
  group_by(producer) %>%
  summarize(
    mean_danceability = mean(danceability, na.rm = TRUE),
    sd_danceability = sd(danceability, na.rm = TRUE),
    min_danceability = min(danceability, na.rm = TRUE),
    max_danceability = max(danceability, na.rm = TRUE),
    n = n()
  ) %>%
  arrange(-mean_danceability)

print(danceability_summary)




#### MODEL 3: MULTICLASS LOGISTIC REGRESSION ###


data_reduced$producer_num <- as.factor(as.numeric(as.factor(data_reduced$producer)))

set.seed(853)


producer_split <- 
  data_reduced |>
  initial_split(strata = producer_num)

producer_train <- training(producer_split)
producer_test <- testing(producer_split)

set.seed(853)
producer_folds <- vfold_cv(producer_train, strata = producer_num)
producer_folds


producer_recipe <- 
  recipe(producer_num ~ ., data = producer_train) %>%
  step_rm(producer)


prep(producer_recipe)

glmnet_spec <- 
  multinom_reg(penalty = tune(), mixture = 1) |>
  set_engine("glmnet")



library(themis)

# Comparing the basic model and the downsampled one
wf_set <-
  workflow_set(
    list(basic = producer_recipe,
         downsampling = producer_recipe |> step_downsample(producer_num)),
    list(glmnet = glmnet_spec)
  )

wf_set

narrower_penalty <- penalty(range = c(-3, 0))

# Register parallel backend
doParallel::registerDoParallel()

set.seed(853)

library(yardstick)


tune_rs <- 
  workflow_map(
    wf_set,
    "tune_grid",
    resamples = producer_folds,
    grid = 15,
    metrics = metric_set(accuracy, mn_log_loss, sens, spec),
    param_info = parameters(narrower_penalty)
  )

tune_rs

autoplot(tune_rs) + theme(legend.position = "none")

rank_results(tune_rs, rank_metric = "sensitivity")
rank_results(tune_rs, rank_metric = "mn_log_loss")
rank_results(tune_rs, rank_metric = "accuracy")
rank_results(tune_rs, rank_metric = "specificity")

downsample_rs <-
  tune_rs %>%
  extract_workflow_set_result("downsampling_glmnet")

# Tuning result for recipe that included downsampling + glmnet model
plotdownsample <- autoplot(downsample_rs) + ggtitle("Downsampled Data")

basic_rs <-
  tune_rs %>%
  extract_workflow_set_result("basic_glmnet")

plotbasic <- autoplot(basic_rs) + ggtitle("Original Data")

library(gridExtra)
grid.arrange(plotdownsample, plotbasic, ncol=2)

best_penalty <- 
  downsample_rs %>%
  select_by_one_std_err(-penalty, metric = "mn_log_loss")

best_penalty

# Last fit is a helper function that fits one time to the training data
# and evaluates one time on the testing data
final_fit <-  
  wf_set %>% 
  extract_workflow("downsampling_glmnet") %>%
  finalize_workflow(best_penalty) %>%
  last_fit(producer_split)

final_fit

collect_metrics(final_fit)

producer_matrix <- collect_predictions(final_fit) %>%
  conf_mat(producer_num, .pred_class)

producer_matrix

producer_matrix_df <- as.data.frame(producer_matrix$table)

# Variable importance
library(vip)
producer_vip <-
  extract_fit_engine(final_fit) %>%
  vi()

producer_vip

producer_vip %>%
  group_by(Sign) %>%
  slice_max(Importance, n = 15) %>%
  ungroup() %>%
  ggplot(aes(Importance, fct_reorder(Variable, Importance), fill = Sign)) + 
  geom_col() +
  facet_wrap(vars(Sign), scales = "free_y") +
  labs(y = NULL) +
  theme(legend.position = "none")
